ParticleEngine --> ParticleSystem All Subclasses of StaticEntity now take floats as width/height instead of int. Added possibility to set custom BlendFunction on Sprites.

+++ src/org/anddev/andengine/entity/DynamicEntity.java
@@ -25,7 +25,7 @@ public abstract class DynamicEntity extends StaticEntity {
 	// Constructors
 	// ===========================================================
 
-	public DynamicEntity(final float pX, final float pY, final int pWidth, final int pHeight) {
+	public DynamicEntity(final float pX, final float pY, final float pWidth, final float pHeight) {
 		super(pX, pY, pWidth, pHeight);
 	}
